fix: [notification] 유효하지 않은 토큰 삭제 오류 해결
- 배치성 작업에 트랜잭션을 설정하지 않아 TransactionRequiredException 예외가 발생

+++ server/notification-service/src/main/java/com/kiosk/server/service/impl/DeletePushTokensServiceImpl.java
@@ -5,6 +5,7 @@ import com.kiosk.server.domain.PushTokenRepository;
 import com.kiosk.server.service.DeletePushTokensService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -15,7 +16,11 @@ public class DeletePushTokensServiceImpl implements DeletePushTokensService {
     private final PushTokenRepository pushTokenRepository;
 
     @Override
+    @Transactional
     public void doService(List<PushToken> pushTokens) {
+        if (pushTokens == null || pushTokens.isEmpty()) {
+            return;
+        }
         pushTokenRepository.deletePushTokensInBatch(pushTokens);
     }
 }
